Notas CRUD

+++ src/views/Notas.vue
@@ -11,39 +11,62 @@
       @dismiss-count-down="countDownChanged"
     >{{alerta.texto}}
     </b-alert>
+    <!-- Mi buscador -->
+    <section class="d-flex flex-row-reverse form-inline">
+          <b-input-group prepend="Buscar" class="mb-12 mr-sm-12 mb-sm-0">
+            <b-input id="buscar" placeholder="Título o descripción" v-model="busqueda"></b-input>
+          </b-input-group>
+    </section>
     <!-- Agregamos el formuario de agregar nota, antes un botón que muestre y oculte el formulario -->
     <!-- https://bootstrap-vue.org/docs/components/form/#form -->
-    <section class="d-flex flex-row">
-      <b-button class="btn-success btn-sm mx-2" @click="formAgregar=!formAgregar">
-          <b-icon icon="file-plus"></b-icon> Nueva
-      </b-button>
-    </section>
-    <b-form @submit.prevent="formAceptar" @reset="formLimpiar" v-if="formAgregar">
-      <h4 class="text-center">Agregar nueva Nota</h4>
-      <b-form-group id="input-titulo">
-        <b-form-input id="titulo" type="text" placeholder="Título de la nota" required v-model="nota.titulo" ></b-form-input>
-      </b-form-group>
-      <b-form-group id="input-des">
-        <b-form-input id="descripcion" type="text" placeholder="Descripción de la nota" required v-model="nota.descripcion" ></b-form-input>
-      </b-form-group>
-      <b-form-group id="input-botones">
-        <b-button type="submit" variant="primary">Agregar</b-button>&nbsp;&nbsp;
-        <b-button type="reset" variant="danger">Cancelar</b-button>
-      </b-form-group>
-    </b-form>
+    <template v-if="!formEditar"> <!-- Oculto todo si estoy en agregar -->
+      <section class="d-flex flex-row">
+        <b-button class="btn-success btn-sm mx-2" @click="formAgregar=!formAgregar">
+            <b-icon icon="file-plus"></b-icon> Nueva
+        </b-button>
+      </section>
+      <!-- Formulario de nueva nota -->
+      <b-form @submit.prevent="formAceptarNueva" @reset="formLimpiarNueva" v-if="formAgregar">
+        <h4 class="text-center">Agregar nueva Nota</h4>
+        <b-form-group id="input-titulo">
+          <b-form-input id="titulo" type="text" placeholder="Título de la nota" required v-model="nota.titulo" ></b-form-input>
+        </b-form-group>
+        <b-form-group id="input-des">
+          <b-form-input id="descripcion" type="text" placeholder="Descripción de la nota" required v-model="nota.descripcion" ></b-form-input>
+        </b-form-group>
+        <b-form-group id="input-botones">
+          <b-button type="submit" variant="primary mx-2">Agregar</b-button>
+          <b-button type="reset" variant="danger mx-2">Cancelar</b-button>
+        </b-form-group>
+      </b-form>
+    </template>
+    <!-- Formulario de editar nota -->
+      <b-form @submit.prevent="formAceptarEditar" @reset="formEditar=false" v-if="formEditar">
+        <h4 class="text-center">Editar Nota</h4>
+        <b-form-group id="input-titulo">
+          <b-form-input id="titulo" type="text" placeholder="Título de la nota" required v-model="nota.titulo" ></b-form-input>
+        </b-form-group>
+        <b-form-group id="input-des">
+          <b-form-input id="descripcion" type="text" placeholder="Descripción de la nota" required v-model="nota.descripcion" ></b-form-input>
+        </b-form-group>
+        <b-form-group id="input-botones">
+          <b-button type="submit" variant="warning mx-2">Modificar</b-button>
+          <b-button type="reset" variant="danger mx-2">Cancelar</b-button>
+        </b-form-group>
+      </b-form>
     <hr>
     <!-- Pintamos la tabla del componente boosrapt b-table: https://bootstrap-vue.org/docs/components/table -->
-    <b-table striped responsive hover :items="notas" :fields="tablaEncabezados">
+    <b-table striped responsive hover :items="filtroNotas" :fields="tablaEncabezados">
       <!-- Por cad acelda indicamos como queremos que se renderice y qué campos -->
       <template v-slot:cell(#)="row">{{row.item._id}}</template>
       <template v-slot:cell(titulo)="row">{{row.item.titulo}}</template>
       <template v-slot:cell(descripcion)="row">{{row.item.descripcion}}</template>
       <template v-slot:cell(fecha)="row">{{row.item.fecha | moment("dddd, D MMMM YYYY, HH:mm:ss")}}</template>
       <template v-slot:cell(acciones)="row">
-        <b-button class="btn-warning btn-sm mx-2" @click="activarEdicion(row.item._id)">
+        <b-button class="btn-warning btn-sm mx-2 my-1" @click="activarEdicion(row.item._id)">
           <b-icon icon="pencil-square" aria-hidden="true"></b-icon> Actualizar
         </b-button>
-        <b-button class="btn-danger btn-sm mx-2" @click="eliminarNota(row.item._id)">
+        <b-button class="btn-danger btn-sm mx-2 my-1" @click="eliminarNota(row.item._id)">
           <b-icon icon="trash"></b-icon> Eliminar
         </b-button>
       </template>
@@ -52,6 +75,8 @@
 </template>
 
 <script>
+/* eslint-disable no-restricted-globals */
+
 import NotasService from '../services/NotasService';
 
 export default {
@@ -72,49 +97,73 @@ export default {
       ],
       // Para la alerta
       alerta: { color: 'success', texto: '' },
-      dismissSecs: 5,
+      dismissSecs: 3,
       dismissCountDown: 0,
       // Muestra el formulario de agregar nota
       formAgregar: false,
+      // Mustra elf ormulario de editar una nota
+      formEditar: false,
+      // Para buscar
+      busqueda: '',
     };
   },
 
   // Al crearme
   created() {
-    // Consultamos todas las notas
-    NotasService.get().then((notas) => {
-      this.notas = notas.data;
-    });
+    // cargamos las notas
+    this.cargarNotas();
+  },
+  // campos computados
+  computed: {
+    // Para manear el filtro
+    filtroNotas() {
+      return this.notas.filter((nota) => nota.titulo.toLowerCase().includes(this.busqueda.toLowerCase()) || nota.descripcion.toLowerCase().includes(this.busqueda.toLowerCase()));
+    },
   },
 
   // Mi métodos
   methods: {
     // activa la edición.
     activarEdicion(id) {
-      console.log(`Editar: ${id}`);
+      this.formEditar = true;
+      // cojemos la nota
+      // si lo hago así, el problema es la reactividad, se ve los cambios en la tabla directamente
+      // eslint-disable-next-line no-underscore-dangle
+      // this.nota = this.notas.find((n) => n._id === id);
+      // La consultamos del servicio, así praticamos
+      NotasService.getById(id)
+        .then((res) => {
+          this.nota = res.data;
+        })
+        .catch((e) => {
+          console.log(e.response);
+        });
     },
     // elimina una nota
     eliminarNota(id) {
-      NotasService.delete(id)
+      // eslint-disable-next-line no-alert
+      if (confirm('¿Está seguro?')) {
+        NotasService.delete(id)
         // Si va bien
-        .then((res) => {
+          .then((res) => {
           // Elimino del array
           // eslint-disable-next-line no-underscore-dangle
-          const index = this.notas.findIndex((item) => item._id === res.data._id);
-          this.notas.splice(index, 1);
-          // Alerta de mensaje
-          this.alerta.texto = '¡Nota eliminada!';
-          this.alerta.color = 'danger';
-          this.showAlert();
-        })
+            const index = this.notas.findIndex((item) => item._id === res.data._id);
+            this.notas.splice(index, 1);
+            // Alerta de mensaje
+            this.alerta.texto = '¡Nota eliminada!';
+            this.alerta.color = 'danger';
+            this.showAlert();
+          })
         // Si falla
-        .catch((e) => {
-          console.log(e.response.data.error.errors.nombre.message);
-          // Alerta de mensaje
-          this.alerta.texto = 'No se ha podido eliminar la nota';
-          this.alerta.color = 'danger';
-          this.showAlert();
-        });
+          .catch((e) => {
+            console.log(e.response);
+            // Alerta de mensaje
+            this.alerta.texto = 'No se ha podido eliminar la nota';
+            this.alerta.color = 'danger';
+            this.showAlert();
+          });
+      }
     },
     // agrega una nueva nota
     agregarNota() {
@@ -130,13 +179,63 @@ export default {
         })
         // Si falla
         .catch((e) => {
-          console.log(e.response.data.error.errors.nombre.message);
+          console.log(e.response);
           // Alerta de mensaje
           this.alerta.texto = 'No se ha podido insertar la nota';
           this.alerta.color = 'danger';
           this.showAlert();
         });
+      this.formAgregar = false;
       this.nota = {};
+    },
+    // edita una nueva nota
+    editarNota() {
+      // eslint-disable-next-line no-underscore-dangle
+      NotasService.put(this.nota._id, this.nota)
+        .then((res) => {
+          // Cambialos los datos en la tabla (podríamos ahorranos esto cargando la tabla directamente con el serviio,
+          // pero es mas rapido así)
+          console.log(res);
+          // eslint-disable-next-line no-underscore-dangle
+          // eslint-disable-next-line no-underscore-dangle
+          const notaMod = this.notas.find((n) => (n._id === this.nota._id));
+          // Solo añadimos los campos que modifico, los otros son iguales
+          notaMod.titulo = this.nota.titulo;
+          notaMod.descripcion = this.nota.descripcion;
+          // Alerta de mensaje
+          this.alerta.texto = 'Nota actualizada';
+          this.alerta.color = 'success';
+          this.showAlert();
+          this.nota = {};
+        })
+        .catch((e) => {
+          console.log(e.response);
+          // Alerta de mensaje
+          this.alerta.texto = 'No se ha podido modificar la nota';
+          this.alerta.color = 'danger';
+          this.showAlert();
+          this.nota = {};
+        });
+      // Ocultamos y limpiamos
+      this.formEditar = false;
+    },
+    // Carga la lista de notas
+    cargarNotas() {
+      this.alerta.texto = 'Cargando notas...';
+      this.alerta.color = 'info';
+      this.showAlert();
+      // Consultamos todas las notas
+      NotasService.get()
+        .then((notas) => {
+          this.notas = notas.data;
+        })
+        .catch((e) => {
+          // Alerta de mensaje
+          console.log(e.response.data.error.errors.nombre.message);
+          this.alerta.texto = 'No se han podido cargar las notas desde el servidor';
+          this.alerta.color = 'danger';
+          this.showAlert();
+        });
     },
     // Metodos de la alerta
     countDownChanged(dismissCountDown) {
@@ -146,13 +245,16 @@ export default {
       this.dismissCountDown = this.dismissSecs;
     },
     // Eventos del formulario
-    formAceptar() {
+    formAceptarNueva() {
       this.agregarNota();
     },
-    formLimpiar() {
+    formLimpiarNueva() {
       // Limpiamos los datos y aceptamos
       this.nota = {};
       this.formAgregar = false;
+    },
+    formAceptarEditar() {
+      this.editarNota();
     },
   },
 };
